+++ frontend/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import './globals.css';
-import { RootProvider } from './RootProvider';
+import { RootProvider } from './providers';
 
 const inter = Inter({ subsets: ['latin'] });
