+++ frontend/src/app/providers/index.tsx
@@ -4,8 +4,8 @@ import client from '@/lib/client';
 import { ApolloProvider } from '@apollo/client';
 import { ThemeProvider } from 'next-themes';
 import { Toaster } from 'sonner';
-import { AuthProvider } from './AuthProvider';
-import RootLayout from './RootLayout';
+import { AuthProvider } from '../AuthProvider';
+import RootLayout from '../RootLayout';
 
 export const RootProvider = ({ children }) => {
   return (
